feat(studio): add support for filtering on files, add unified asset preview component
